refactor(core-interactive-icon): seperate icons into their own folders
