+++ packages/InteractiveIcon/svgs/InteractiveButton/Subtract.jsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react'
 
-import InteractiveIcon, { StyledInteractiveIconSVG } from '../InteractiveButton'
+import InteractiveIcon, { StyledInteractiveIconSVG } from '../../InteractiveButton'
 
 const Subtract = forwardRef((props, ref) => (
   <InteractiveIcon {...props} ref={ref}>
