+++ packages/InteractiveIcon/svgs/InteractiveButton/index.js
@@ -1,0 +1,4 @@
+export { default as Add } from './Add'
+export { default as Close } from './Close'
+export { default as PlayVideo } from './PlayVideo'
+export { default as Subtract } from './Subtract'
