+++ packages/InteractiveIcon/svgs/Limited/ChevronRight.jsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react'
 
-import LimitedInteractiveIcon, { StyledLimitedInteractiveIconSVG } from '../Limited'
+import LimitedInteractiveIcon, { StyledLimitedInteractiveIconSVG } from '../../Limited'
 
 const ChevronRight = forwardRef((props, ref) => (
   <LimitedInteractiveIcon {...props} ref={ref}>
