+++ packages/InteractiveIcon/svgs/Limited/index.js
@@ -1,0 +1,4 @@
+export { default as CaretUp } from './CaretUp'
+export { default as CaretDown } from './CaretDown'
+export { default as ChevronLeft } from './ChevronLeft'
+export { default as ChevronRight } from './ChevronRight'
