+++ packages/InteractiveIcon/svgs/index.js
@@ -1,8 +1,3 @@
-export { default as Add } from './Add'
-export { default as Close } from './Close'
-export { default as PlayVideo } from './PlayVideo'
-export { default as Subtract } from './Subtract'
-export { default as CaretUp } from './CaretUp'
-export { default as CaretDown } from './CaretDown'
-export { default as ChevronLeft } from './ChevronLeft'
-export { default as ChevronRight } from './ChevronRight'
+export * from './InteractiveButton'
+export * from './Limited'
+export * from './NavButton'
